Implemented P5 changes
- Updated DengFOC.h to remove dual motor references, focusing on single dial functionality.
- Modified Dial.cpp and Dial.h to streamline methods for a single dial, removing motor index dependency.
- Enhanced motor_id_calibration.py to reflect changes from motor to dial terminology, updating comments and variable names accordingly.
- Adjusted platformio.ini to change upload port from COM3 to COM4.
- Updated main.cpp to support new dial configuration, including persistent dial ID storage and telemetry adjustments.
- Improved angle and torque calculations to ensure consistency with new dial logic.

+++ lib/dengfoc/AS5600.h
@@ -7,7 +7,7 @@ public:
   Sensor_AS5600(int Mot_Num);
   void Sensor_init(TwoWire *_wire = &Wire);
   void Sensor_update();
-  float getAngle();
+  double getAngle();
   float getVelocity();
   float getMechanicalAngle();
   float getSensorAngle();
@@ -17,9 +17,9 @@ private:
   // AS5600 变量定义
   // int sensor_direction=1;       //编码器旋转方向定义
   float angle_prev = 0;           // 最后一次调用 getSensorAngle() 的输出结果，用于得到完整的圈数和速度
-  long angle_prev_ts = 0;         // 上次调用 getAngle 的时间戳
+  uint32_t angle_prev_ts = 0;     // 上次调用 getAngle 的时间戳
   float vel_angle_prev = 0;       // 最后一次调用 getVelocity 时的角度
-  long vel_angle_prev_ts = 0;     // 最后速度计算时间戳
+  uint32_t vel_angle_prev_ts = 0; // 最后速度计算时间戳
   int32_t full_rotations = 0;     // 总圈数计数
   int32_t vel_full_rotations = 0; // 用于速度计算的先前完整旋转圈数
   float last_valid_angle = 0;     // Last successful sensor reading (returned on I2C error)
